requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==3.2.0
certifi==2025.11.12
click==8.1.8
cloudinary==1.44.0
dnspython==2.7.0
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.122.0
fastapi-cli==0.0.16
fastapi-mail==1.6.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
mysqlclient==2.2.7
pydantic>=2.12.5,<3.0.0
pydantic_core==2.41.5
PyJWT==2.0.1
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
rich==14.2.0
sentry-sdk==2.46.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
argon2-cffi==23.1.0
httpx
